Updated headers to 145286

+++ FactoryGame/Source/FactoryGame/Public/FGResourceSinkSubsystem.h
@@ -65,7 +65,7 @@ public:
 	
 	/** Returns the number of points left until you receive the next coupon */
 	UFUNCTION( BlueprintPure, Category = "FactoryGame|ResourceSink" )
-	int32 GetNumPointsToNextCoupon() const;
+	int64 GetNumPointsToNextCoupon() const;
 	
 	/** Returns the percentage (normalized value) of your progression towards the next coupon */
 	UFUNCTION( BlueprintPure, Category = "FactoryGame|ResourceSink" )
